fix: Clippy warnings (code quality improvements)
- Replace or_insert_with(Vec::new) with or_default()
- Fix unused variable bindings and loop indexing patterns
- Remove absurd comparisons (>= 0 on usize)
- Improve code clarity per clippy lints

+++ crates/chematic-chem/src/descriptors.rs
@@ -1623,7 +1623,7 @@ pub fn hall_kier_alpha(mol: &Molecule) -> f64 {
             17 => 1.02,  // Cl
             35 => 1.20,  // Br
             53 => 1.39,  // I
-            _ => (an * 0.15) as f64, // fallback
+            _ => an * 0.15, // fallback
         };
 
         // Hall-Kier alpha value proportional to radius and degree
